Added main menu, changed file structure, added gamepad support
