--- GamepadHandler.js
@@ -1,20 +1,0 @@
-export default class GamepadHandler {
-    constructor() {
-        this.gamepadIndex = null;
-        this.axes = [0, 0, 0, 0];
-        this.buttons = { RT: false };
-        
-        window.addEventListener("gamepadconnected", (e) => {
-            this.gamepadIndex = e.gamepad.index;
-        });
-    }
-
-    update() {
-        if (this.gamepadIndex === null) return;
-        const gp = navigator.getGamepads()[this.gamepadIndex];
-        if (!gp) return;
-
-        this.axes = [gp.axes[0], gp.axes[1], gp.axes[2], gp.axes[3]];
-        this.buttons.RT = gp.buttons[7].pressed;
-    }
-}
