+++ index.html
@@ -19,6 +19,6 @@
 </head>
 <body>
     <canvas id="gameCanvas"></canvas>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="src/main.js"></script>
 </body>
 </html>
